Add beginning of AudioBufferSource support
The idea is that an AudioBufferSourceNode will have a URL parameter.  A new module - the ResourceLoader - attempts to load all the such resources mentioned in the audiograph and returns either an error or a map of the URL strings to the decoded buffers and this is made available to the assembler.