Introduced a Compiler
This will perform semantic checks after parsing.  It is necessary to check that connections are to ids of nodes that exist but this needs to be done now AFTER parsing completes because we should be allowed to refer to nodes both forwards and backwards (to allow feedback loops in web-audio).  At the moment, the parser only allows backward references.